WS-024: Horizontal menu appear-disappear
